backend/requirements.txt
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2==2.9.9
pyOpenSSL==24.2.1
redis==5.0.7
sqlparse==0.5.0
django-prometheus==2.3.1
pytest
requests